+++ src/main/java/org/sbml4j/repository/sbml/SBMLSpeciesRepository.java
@@ -11,7 +11,7 @@
  * 
  * For a full list of authors, please refer to the file AUTHORS.
  */
-package org.sbml4j.repository.common;
+package org.sbml4j.repository.sbml;
 
 import java.util.List;
