+++ src/main/java/org/sbml4j/service/SimpleSBML/SBMLCompartmentService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.sbml4j.model.common.GraphBaseEntity;
 import org.sbml4j.model.sbml.SBMLCompartment;
-import org.sbml4j.repository.common.SBMLCompartmentRepository;
+import org.sbml4j.repository.sbml.SBMLCompartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
