+++ src/main/java/org/sbml4j/service/SimpleSBML/SBMLSBaseEntityService.java
@@ -2,7 +2,7 @@ package org.sbml4j.service.SimpleSBML;
 
 import org.sbml4j.model.common.GraphBaseEntity;
 import org.sbml4j.model.sbml.SBMLSBaseEntity;
-import org.sbml4j.repository.common.SBMLSBaseEntityRepository;
+import org.sbml4j.repository.sbml.SBMLSBaseEntityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
